Calculate phylogenetic community metrics for raster data

Calculate species richness, phylogenetic diversity, evolutionary distinctiveness, phylogenetic endemism and weighted endemism using rasters as input

Usage

.rast.geo.phylo(
  x,
  inv.R,
  edge.path,
  branch.length,
  n.descen,
  filename = "",
  ...
)

Arguments

x

SpatRaster. A SpatRaster containing presence-absence data (0 or 1) for a set of species. The layers (species) must be sorted according to the tree order. See the phylo.pres function.

inv.R

SpatRaster. inverse of range size calculated from inv.range

edge.path

matrix representing the paths through the tree from root to each tip. See phylo.pres

branch.length

numeric. A named numerical vector containing the branch length for each species.

n.descen

numeric. A Named numeric vector of number of descendants for each branch

filename

character. Output filename

...

additional arguments passed for terra::app

Value

SpatRaster with one layer for each metric

Details

Community metrics calculated:

Phylogenetic diversity (Faith 1992)

Species Richness

Evolutionary distinctiveness by fair-proportion (Isaac et al. 2007)

Phylogenetic endemism (Rosauer et al. 2009)

Weighted endemism (Crisp et al. 2001, Williams et al. 1994)

References

Rosauer, D. A. N., Laffan, S. W., Crisp, M. D., Donnellan, S. C. and Cook, L. G. (2009). Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular ecology, 18(19), 4061-4072.

Faith, D. P. (1992). Conservation evaluation and phylogenetic diversity. Biological conservation, 61(1), 1-10.

Williams, P.H., Humphries, C.J., Forey, P.L., Humphries, C.J. and VaneWright, R.I. (1994). Biodiversity, taxonomic relatedness, and endemism in conservation. In: Systematics and Conservation Evaluation (eds Forey PL, Humphries CJ, Vane-Wright RI), p. 438. Oxford University Press, Oxford.

Crisp, M., Laffan, S., Linder, H. and Monro, A. (2001). Endemism in theAustralian flora. Journal of Biogeography, 28, 183–198.

Isaac, N. J., Turvey, S. T., Collen, B., Waterman, C. and Baillie, J. E. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2, e296.

Laffan, S. W., Rosauer, D. F., Di Virgilio, G., Miller, J. T., González‐Orozco, C. E., Knerr, N., ... & Mishler, B. D. (2016). Range‐weighted metrics of species and phylogenetic turnover can better resolve biogeographic transition zones. Methods in Ecology and Evolution, 7(5), 580-588.

Author

Neander Marcel Heming